+++ CHANGES.rst
@@ -22,6 +22,7 @@ Release History
 0.3.2 (unreleased)
 ==================
 
+- Token-based authentication
 - Dual-stack IPv4/IPv6 support
 
 0.3.1 (October 17, 2017)